Add: HorizontalLabeledSlider component
